Move Painting to base (gallery) module to domain package
- the domain model of Painting contains more information than is necessary for the UI on some screens
- we could introduce a UiModel which contains only the pertinent information for those screens
