+++ app/src/main/java/com/ataulm/artcollector/domain/Painting.kt
@@ -1,11 +1,11 @@
-package com.ataulm.artcollector.gallery.domain
+package com.ataulm.artcollector.domain
 
-import com.ataulm.artcollector.domain.Artist
-
-internal data class Painting(
+data class Painting(
         val id: String,
         val title: String,
+        val webUrl: String,
         val description: String?,
+        val creditLine: String?,
         val imageUrl: String,
         val artist: Artist
 )
