+++ app/src/main/java/com/ataulm/artcollector/gallery/domain/Gallery.kt
@@ -1,3 +1,5 @@
 package com.ataulm.artcollector.gallery.domain
 
+import com.ataulm.artcollector.domain.Painting
+
 internal class Gallery(collection: Collection<Painting>) : ArrayList<Painting>(collection)
