+++ app/src/main/java/com/ataulm/artcollector/gallery/ui/GalleryViewModel.kt
@@ -7,7 +7,7 @@ import com.ataulm.artcollector.Event
 import com.ataulm.artcollector.domain.Artist
 import com.ataulm.artcollector.gallery.domain.Gallery
 import com.ataulm.artcollector.gallery.domain.GetGalleryUseCase
-import com.ataulm.artcollector.gallery.domain.Painting
+import com.ataulm.artcollector.domain.Painting
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.Job
