+++ painting/src/main/java/com/ataulm/artcollector/painting/domain/PaintingRepository.kt
@@ -1,5 +1,7 @@
 package com.ataulm.artcollector.painting.domain
 
+import com.ataulm.artcollector.domain.Painting
+
 internal interface PaintingRepository {
 
     suspend fun painting(): Painting
