+++ painting/src/main/java/com/ataulm/artcollector/painting/ui/PaintingActivity.kt
@@ -6,9 +6,9 @@ import android.os.Bundle
 import android.view.View
 import androidx.appcompat.app.AppCompatActivity
 import com.ataulm.artcollector.DataObserver
+import com.ataulm.artcollector.domain.Painting
 import com.ataulm.artcollector.imageUrl
 import com.ataulm.artcollector.painting.R
-import com.ataulm.artcollector.painting.domain.Painting
 import com.ataulm.artcollector.painting.domain.PaintingId
 import com.ataulm.artcollector.painting.injectDependencies
 import com.ataulm.artcollector.webIntent
